Swap out JSON for BSON.

+++ leafset.go
@@ -33,7 +33,7 @@ func (l *leafSet) insertNode(node Node) (*Node, error) {
 	return l.insertValues(node.ID, node.LocalIP, node.GlobalIP, node.Region, node.Port, node.routingTableVersion, node.leafsetVersion, node.neighborhoodSetVersion)
 }
 
-func (l *leafSet) insertValues(id NodeID, localIP, globalIP, region string, port int, rTVersion, lSVersion, nSVersion uint64) (*Node, error) {
+func (l *leafSet) insertValues(id NodeID, localIP, globalIP, region string, port int, rTVersion, lSVersion, nSVersion uint32) (*Node, error) {
 	l.lock.Lock()
 	defer l.lock.Unlock()
 	node := NewNode(id, localIP, globalIP, region, port)
